Almost finished Biofilm  productivity code
I did everything but calculate the adjusted ys and make the figure

The results are slightly different than the previous version. More thing are showing up in the best model set (pH and DOC) so I want to go back and double check that the data are accurate by comparing datasets to what I had before.